Redesignet total measurement graph to be temperature instead
